pyproject.toml
[project]
name = "itk-academy-testovoe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "anyio>=4.12.1",
    "asgi-lifespan>=2.1.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.129.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-env>=1.3.2",
    "python-dotenv>=1.2.1",
    "ruff>=0.15.1",
    "sqlalchemy[asyncio]>=2.0.46",
    "uvicorn>=0.40.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"


[tool.pytest_env]
ENV = "test"
POSTGRES_USER= "test"
POSTGRES_PASSWORD= "test"
POSTGRES_HOST = "test-db"

[tool.ruff]
line-length = 88

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.ruff.lint]
extend-select = ["E", "W"]

